+++ src/aui-scheduler/js/aui-scheduler-base-view.js
@@ -241,6 +241,20 @@ var SchedulerView = A.Component.create({
         flushViewCache: function() {},
 
         /**
+         * Returns the date interval in which this view shows events for.
+         *
+         * @method getDateInterval
+         * @return {Object} Object with 2 keys: startDate and endDate. Undefined
+         *   keys are interpreted as unlimited sides of the interval.
+         */
+        getDateInterval: function() {
+            return {
+                endDate: DateMath.toLastHour(DateMath.subtract(this.getNextDate(), DateMath.DAY, 1)),
+                startDate: this.getAdjustedViewDate(this.get('scheduler').get('viewDate'))
+            };
+        },
+
+        /**
          * TODO. Wanna help? Please send a Pull Request.
          *
          * @method getNextDate